// frontend/src/api/models/PositionRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Clinic Management API
 * API documentation for the Clinic Management system
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PositionRequest
 */
export interface PositionRequest {
    /**
     * 
     * @type {string}
     * @memberof PositionRequest
     */
    positionCode?: string;
    /**
     * 
     * @type {string}
     * @memberof PositionRequest
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof PositionRequest
     */
    description?: string;
}

/**
 * Check if a given object implements the PositionRequest interface.
 */
export function instanceOfPositionRequest(value: object): value is PositionRequest {
    return true;
}

export function PositionRequestFromJSON(json: any): PositionRequest {
    return PositionRequestFromJSONTyped(json, false);
}

export function PositionRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): PositionRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'positionCode': json['positionCode'] == null ? undefined : json['positionCode'],
        'title': json['title'] == null ? undefined : json['title'],
        'description': json['description'] == null ? undefined : json['description'],
    };
}

export function PositionRequestToJSON(json: any): PositionRequest {
    return PositionRequestToJSONTyped(json, false);
}

export function PositionRequestToJSONTyped(value?: PositionRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'positionCode': value['positionCode'],
        'title': value['title'],
        'description': value['description'],
    };
}