// frontend/src/api/models/PositionResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Clinic Management API
 * API documentation for the Clinic Management system
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PositionResponse
 */
export interface PositionResponse {
    /**
     * 
     * @type {string}
     * @memberof PositionResponse
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof PositionResponse
     */
    positionCode?: string;
    /**
     * 
     * @type {string}
     * @memberof PositionResponse
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof PositionResponse
     */
    description?: string;
}

/**
 * Check if a given object implements the PositionResponse interface.
 */
export function instanceOfPositionResponse(value: object): value is PositionResponse {
    return true;
}

export function PositionResponseFromJSON(json: any): PositionResponse {
    return PositionResponseFromJSONTyped(json, false);
}

export function PositionResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): PositionResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'positionCode': json['positionCode'] == null ? undefined : json['positionCode'],
        'title': json['title'] == null ? undefined : json['title'],
        'description': json['description'] == null ? undefined : json['description'],
    };
}

export function PositionResponseToJSON(json: any): PositionResponse {
    return PositionResponseToJSONTyped(json, false);
}

export function PositionResponseToJSONTyped(value?: PositionResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'positionCode': value['positionCode'],
        'title': value['title'],
        'description': value['description'],
    };
}